test: split request day rollup coverage
